fix lot of bugs 2

+++ zenovis/include/zenovis/RenderEngine.h
@@ -25,8 +25,8 @@ struct RenderEngine {
     virtual void cleanupWhenExit() = 0;
 
     virtual ~RenderEngine() = default;
-    virtual std::optional<glm::vec3> getClickedPos(int x, int y) { return {}; }
-    virtual std::optional<std::tuple<std::string, uint32_t, uint32_t>> getClickedId(int x, int y) { return {}; }
+    virtual std::optional<glm::vec3> getClickedPos(float x, float y) { return {}; }
+    virtual std::optional<std::tuple<std::string, uint32_t, uint32_t>> getClickedId(float x, float y) { return {}; }
     virtual void load_matrix_objects(std::vector<std::shared_ptr<zeno::IObject>> matrixs) {};
     virtual void outlineInit(Json const &msg) {};
